+++ Bluejay/Bluejay/DisconnectionResult.swift
@@ -7,14 +7,11 @@
 //
 
 import Foundation
-import CoreBluetooth
 
 /// Indicates a successful, cancelled, or failed disconnection attempt, where the success case contains the peripheral disconnected from.
 public enum DisconnectionResult {
     /// The disconnection is successful, and the disconnected peripheral is captured in the associated value.
-    case success(CBPeripheral)
-    /// The disconnection is cancelled for a reason.
-    case cancelled
+    case disconnected(Peripheral)
     /// The disconnection has failed unexpectedly with an error.
     case failure(Error)
 }