+++ Bluejay/Bluejay/DiscoverService.swift
@@ -60,20 +60,7 @@ class DiscoverService: Operation {
         
         updateQueue()
     }
-    
-    func cancel() {
-        cancelled()
-    }
-    
-    func cancelled() {
-        state = .cancelled
-        
-        callback?(.cancelled)
-        callback = nil
-        
-        updateQueue()
-    }
-    
+            
     func fail(_ error: Error) {
         state = .failed(error)