+++ Bluejay/Bluejay/WriteCharacteristic.swift
@@ -82,22 +82,7 @@ class WriteCharacteristic<T: Sendable>: Operation {
             preconditionFailure("Expecting write to characteristic: \(characteristicIdentifier.uuid), but received event: \(event)")
         }
     }
-    
-    func cancel() {
-        cancelled()
-    }
-    
-    func cancelled() {
-        state = .cancelled
-        
-        log("Cancelled write to \(characteristicIdentifier.uuid) on \(peripheral.identifier).")
-        
-        callback?(.cancelled)
-        callback = nil
-        
-        updateQueue()
-    }
-    
+            
     func fail(_ error: Error) {
         state = .failed(error)